refactor(unused_deps_py): Move library definition to its own build file
